Add gnuplot warn

+++ examples/plot-dsp.sh
@@ -13,6 +13,8 @@ STREAM=outputAll
 
 if ! xcompiler -q query-dsp.rql ; then exit 1 ; fi
 
+if ! which gnuplot ; then echo "install gnuplot!" ; exit 1 ; fi
+
 rm nohup.out
 nohup xretractor &
